refactor/good: Move getGoods logic to model
